Generalize ListPanelDialog so it has a scrollview rather than a recyclerview
Before the commit, we have a Recylerview in ListPanelDialog. This is a very hard
requirement which does not give us the elasticity to change the inner list
implementation. The Recyclerview is replaced by a NestedScrollView container
so the Fragements or Views can be inserted. The implementation of the inner list
is thus a responsibility of the inner Fragment/View.