Implement PanelFragment as a transaction to close ListPanelDialog from childrens
We need an interface to provide children fragments of ListPanelDialog to close
the ListPanelDialog when necessary. We ensure that all childrens are instance
of PanelFragment, and provide a closePanel() for PanelFragment instances

BrowsingHistoryFragment is one that needs this and is setup to close panel with
the aforementioned methodology in this commit.