file-pick: to handle file choose action
To copy some logic from Fennec to provide file choose intent.

In Fennec we create an intent for file uploading, and the intent could
also launch Camera to take picture. We don't do it in Zerda for now to
make things eaiser.

We have to check external storage reading permission before launch any
activity for file picking, otherwise we might get exception which cannot
be caught.