implement Receiver and Service for file moving
DownloadCompleteReceiver is a receiver to listen Download-complete
broadcast, it will start RelocateService in another thread to perform
advance file operation.

RelocateService is a service running on working thread, to perform
advanced file operation for downloaded file - to relocate file to
respect user's preferences.

If the relocating success, the serivce will also update database.