ux: change wording from 'screenshots' to 'my shots'
We found put 'screenshots' in bottom sheet often confuse users,
therefore we change wording to 'my shots'.

However we used 'screenshot' in code base widely, include drawables name
and methods name. It is hard to decide whether to replace all of them or
not.

Consider that we might change wording again, this patch just replace
things as less as possible.