workaround: to around TransactionTooLargeException
refer to: https://issuetracker.google.com/issues/37103380

Rather than a slient warning, Android starts throwing exception since N.
But we have no chance to catch the exception.

It might happens in any place of code, and we guess WebView hit this mostly.
Here we try to workaround it by removing history manually, if it is LARGE.

to fix #1107