tabs: add basic Tabs implementation
This commit pull in some bacis implementation for multi-tabs.
It defines model class, some interfaces as callback. Try to bring
multi-tabs feature to UI in a easier way.

It tries to abstract WebView and hope we can involve other system(ie.
GeckoView) in the future.

Be carefully. This implementation is still under heavy development
and might be changed quickly.