[Widget] Refine TabCounter to prevent layout inflation at runtime
Since the child views in TabCounter have fixed size and layout
position, we don't need to declare the layout in XML and inflate it at
runtime. Instead, we can directly hardcode the size/position value in
java files.

The change removes the layout inflation on main thread at runtime,
which means we can be more confident to measure/layout/draw views in
16ms, even in lower end devices.