Closes #2025: Trim invalid Firebase Event name and add a whitelist for existing pref key
When chaning the share preference, we sent telemetry event with value=<pref key name>

This is a problem for Firebase, cause method+object+value will over the 40 characters limit.

So I make a whitelist to all current pre keys, and tirm them programatically.

Value lenght now is (40 chars - method - object )

In the future, if anyone want to add more pref and exceed Firebase event limit, a crash will be thrown.