Add ktlint
Android Components is using ktlint and detekt for style check and static code
analysis for Kotlin. They align with Google's official Kotlin guidelines
and doesn't have a lot of conflicts in our app. So the effort to add it
is low. If we want to contributte to a-c, we should adopt them as early
as possible.

I haven't make gradle check task depend on them cause I want others to
test it before we make it as part of our shipping standard.

Note: ktlint hass auto formatter support. (Just run ./gradlew ktlintFormat)
detekt doesn't have that since it's not a style check tool. It's a bad
smell checker so only the developer knows how to fix it. (just like
findbugs)

For the reference, visit
[ktlint](https://github.com/shyiko/ktlint)
[detekt](https://github.com/arturbosch/detekt)