Add category for screenshot. closes #1987
We load the category mapping list in the Application class, and use it without
making sure it's there. It's something we need to fix in the future.

Ideally, every config should be ready before we really use it.
We didn't do it here because
1. The schedule doesn't allow us.
2. This feature may be removed in the future.

After confirmed with UX and product, we'll display empty string is the
mapping is not ready.