privately: drop savedInstanceState to reset Navigation
If the Activity is destroyed(turn on Don't Keep Activity), the
Navigation component will remember its current destination and try to
restore it.

We don't save webView state in private-mode for now, if current
destination is in BrowserFragment, then we will get blank screen.

This commit drop the saved-state to reset Navigation.