Send the last ping before the user disable "Send Usage Data" closes #2487
Confirmed with UX and product.
We'll send this ping using our best effort.
If the battery is low and the network is slow, the ping won't be sent.
This is because the Telemetry library uses JobScheduler to upload the ping.
And there's no control at our end. In the above situation, data loss is acceptable.