Use PrivateModeActivty to determine if the application is in private mode.
FocusApplication need to know if we are in private mode process, so it
can know when to override getCacheDir to provide the cache folder for
webview/chromium

We use process's name to see if it contains "private_mode" using
acitivityManager.getRunningAppProcesses(). But this method
sometimes return null list so a NullPointerException will be thrown
if we want to use it to check if we are in private mode.

We want to use a new way to determine if we are in prviate mode
process. We want to use the existance of PrivateModeActivity here.

But Telemetry.init() call in FocsApplication will make this hard cause
it make getCacheDir called much more early than PrivateModeActivity is
created.  This is because it checked the webview's version in
TelemtryWrapper.init() and call the constructor of WebView.

This makes the chromium code call context.getApplicationContext().getCacheDir() too
early even before the acitvity is created.

Our solution here is to remove the webview version check. But maybe we
can put the version check after the first webView is created and save it
to a shared preference. At the next launch, we just read the pref but we
don't create a webview again just to check the version.

The assumption here is that as long as there's PrivateModateActivity,
this application is in Private Mode process.
This assumption won't hold true if private mode leaves in the same
process of normal mode. But this is the desgin right now.

We should also test if the activity is killed by the system, will this
still work?ï¼assume it will cause there'll never be a
PrivateModeActivity in normal mode.
Need more testing so this can be landed.