Auto generate the wiki for telemetry event pings.
We'll aslo generate an untracked csv file for transfering focus-event to Amplitude events.
We only check in the markdown output for wiki, because we want to display the link on current wiki page.
And the markdown output will be in-tree, and there's a taks to check if the .md file is committed.
If not, commpile will fail.