Control ShiftDrawable from AnimantedProgressBar closes #2454
Previously, we use ShiftDrawable's visiblity to control the animator running or not.
Since now BrowserFragment (which host the AnimatedProgressBar) will always visible, the animator will be always running.

Now I use the AnimatedProgressBar  to control the animator and stop the animator when the closing animation is closed.

I tried make it implements Animateable and Runnable, and use scheduleSelf(runnalbe,delay) to replace Animator.
But since animator talks to Choreographer directly, and scheduleSelf will ask it's callback (View.java) to talk to Choreographer,
I'd like to use the first approach.