Try to use Animateable.
Calculating the fraction by ourselves is hard. Cause it's hard to get the playback time correct without knowing the exact time
the first fram is played(consider the situation of frame lost, etc)

Need more time to get the effect in draw() to work.

Other than that, the progress bar is good to work now.