Closes #3079 - Add telemetry pings to better understand Firebase lifecycle
In this commit, I removed a bunch of things.
Basically because in our current architecture, MainActivity contains BrowserFragment
So all the toolbar/UI will be inflated when at the same time as FirstRunFragment.

The only useful checks is when Firstrun ( the real Firstrun, not update first run) finishes.
This is also the timing when we add HomeFragment.

The commit here only implment the minimun effort for the lifecycle check.