Reuse HomeFragment instead of recreating it.
The reason for this commit:
1. Reuse HomeFragment. So we can keep its state after raising BrowserFragment, and coming back.
2. Previously we've used fixed margin to avoid status bar and navigationbar. Now we use WindowsInset
which works better with cut-off, and more flexable.

The second point is the main reason for this refactoring:
HomeFragment needs OnApplyWindowInsetListener to set the propper padding.
And only when calling setSystemUIVisibility in MainActivity will make sure the window inset events gets trigger.
So I need to put HomeFragment in the MainActivity's content view so it can recieve the event.