Using constructor injection for remote config default values in FirebaseImpl
Previously, we can put a json file in external stroage, and FirebaseHelper will read that json file as it's default value.
Instead of read the default value from resource strings and some constants in the code.

Now Remote Config is always on. Doing so will only replace the default value, not the remote config value.

For now on, if anyone want to test remote config value in UI tests or manually, he/she need to change the value in the server.
If he/she want to do unit test, use FirebaseNoOpImp and inject the default value in constructor.