Issue #3440 - Decouple hasEcShoppingLink check from NewFeature and remote config initialization.
Previously, we check "hasEcShoppingLink" in NewFeatureNotice's private constructor.
Now, we only pull the information from Remote Config when we really want to show the UI.

The problem with the former approach, is that NewFeatureNotice could be initialized before
Remote Config has completed its initialization. So upgrade onboarding won't show most of the time.
( hasEcShoppingLink is always false)
Now we get rid of that problem, and defer the check when the UI is about to show.( Always look at
the latest Remote Config value via AppConstantWrapper)

Could there be a timing that Remote Config is not ready when the UI is about to show? Yes, there could be.
But this is less troublesome because we won't show the UI anyway. (Both Onboarding and the feature)

This commit alaso allow us to pass null context into FirebaseHelper.getRcString(). Since we only use
that context when the default string is from resources, and the default value for shopping links is
hardcoded (empty string), we could safely pass null here.