Add Tab for Content Portal
Previously, we use AppConfigWrapper to decide what view to inflate.
Now, we use AppConfigWrapper to decide what features we have and put them in an ArrayList.
Then we pass the ArrayList to ContentFragment. Let Content Fragment decide what to inflate there.

Although we can add ContentFragment directly and decide what to inflate there.
I want to keep the old implementation so we can switch back if we want to.

.