Integrate  Banner in Coupon list
Banner loading has lifecycle.

Previously, it requires HomeFragment to observe the data coming from the banner module.
Now, we use ContentFragment to observe it.

Previous BannerViewModel only stays in HomeFragment. Now it is a single instance in MainActivity
cause fragments (HomeFragment, EcFragment) need to observe it.

After the onChange event in EcFragment, we create the BannerAdapter and TelemetryListener in the
observer and update the recycler view for banner.