Support ViewPager in BottomSheetBehavior
BottomSheetBehavior can only have one scrollingChild at a time.
It's only updated when onLayoutChild by findScrollingChild() method.

The original implmentation of findScrollingChild() will always return the first scrolling child.
This is not enough for ViewPager cause the scrolling child will change to it's current selected child.

findScrollingChild() is package-private and annotated with @Visible for testing. Override it will easily
break in the future. (When we upgrade to Androidx maybe?)

It's better we have custom behaviour. I'll do it in the next commit.