Move Telemetry related output file to generated folder.
I don't know why, but for some reason the old Annotation Processor suddenly didn't do its job.
I figure out that my JAVA_HOME was set to Java 11 when I do server side development.
After switching back to Java 8, kapt can work. But it can't wirte output to the folder outside of generated folder.
I checked the last time writing to external folder worked is July 29, 2019 in events.md
During this period, there's no change in Gradle plugin nor Kotlin plugin.

In short, I can't find the cause for this change.
For now, what I can do is to write the generated doc in the generated folder.
If we really want these files, we need to manually copy them.
In the future, it's better we fix this problem, or have a Gralde task copy those genereated docs to where they were.
But this also need to survey if there's an API to let us know where the path is.

I'll pause the effort here.