Align the way we inject HomeViewModel and its depedencies. closes #4942
If we want to initialize HomeViewModel using the old method, the depedencies won't be ready so it'll choose the zero parameter constructor.
Since we use Dagger to inject HomeViewModel and its depedencies, we should do the same here.

The reason why previous versions won't crash is because HomeFragment will always be visible first.
But if we turn on "Don't keep activity" and resume the app while Tab Tray is previously showing, this trash will happen.