upgrade fcm
perf-monitor has a class `com.firebase.iid.zzbb`. It's version is 20.2.0. and it's final.
But in current FCM library, a class named com.firebse.mssage.zzf needs to implement that class, d8 will fail when running gradle task :app:dexBuilderFocusWebkitDebug

```
D8: Type com.google.firebase.iid.zzbb is referenced as an interface from `com.google.firebase.messaging.zzf`.
```

There are two ways to fix this. Using an older version of perf-monitoring. or use a newer version of FCM.
I choose the latter since we'll need to upgrade it anyway in the future. It's better we deal with it now.